+++ src/main/kotlin/lotto/model/LottoResult.kt
@@ -6,7 +6,7 @@ object LottoResult {
     var third: Int = 0
     var fourth: Int = 0
     var fifth: Int = 0
-    var priceRatio: Float = 0f
+    var priceRatio: Double = 0.0
 
     fun reset() {
         first = 0
@@ -14,6 +14,6 @@ object LottoResult {
         third = 0
         fourth = 0
         fifth = 0
-        priceRatio = 0f
+        priceRatio = 0.0
     }
 }
